browser/requirements.txt
flask==2.3.3
torch==2.0.1
zuko==0.4.0
h5py==3.9.0
numpy==1.24.3
Pillow==10.0.0
requests==2.31.0
matplotlib==3.7.2
umap-learn==0.5.3 